# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPack API Documentation",
        "description": "Documentación completa de la API de NewPack",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get all suppliers",
                "operationId": "918f9fb517e5a684df857bb069f14f72",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "supplier_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppliers list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "operationId": "bf86e79b2ce533aff7283f71fc453add",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers/{supplier}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get a supplier by ID",
                "operationId": "d1e12bec91a2485ef7a723d2943a33c7",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update a supplier",
                "operationId": "00e3c737eb0bf0ed7c9175404ef418e4",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Soft delete a supplier",
                "operationId": "8837dedd53289fc9459fa7781c026bd2",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreSupplierRequest": {
                "required": [
                    "company_name"
                ],
                "properties": {
                    "company_name": {
                        "type": "string",
                        "example": "Proveedora SA de CV"
                    },
                    "contact_name": {
                        "type": "string",
                        "example": "Juan Pérez",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "contacto@proveedora.com",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phone_secondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburb_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Industrial",
                        "nullable": true
                    },
                    "exterior_number": {
                        "type": "string",
                        "example": "500",
                        "nullable": true
                    },
                    "interior_number": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "address_reference": {
                        "type": "string",
                        "example": "Zona industrial norte",
                        "nullable": true
                    },
                    "rfc": {
                        "type": "string",
                        "example": "PRO850101AB2",
                        "nullable": true
                    },
                    "legal_name": {
                        "type": "string",
                        "example": "Proveedora SA de CV",
                        "nullable": true
                    },
                    "tax_system": {
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "use_cfdi": {
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "supplier_type": {
                        "type": "string",
                        "enum": [
                            "product",
                            "service",
                            "both"
                        ],
                        "example": "product"
                    },
                    "payment_terms": {
                        "type": "string",
                        "example": "30 días",
                        "nullable": true
                    },
                    "credit_limit": {
                        "type": "number",
                        "format": "float",
                        "example": 100000,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Proveedor confiable",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateSupplierRequest": {
                "properties": {
                    "company_name": {
                        "type": "string",
                        "example": "Proveedora SA de CV"
                    },
                    "contact_name": {
                        "type": "string",
                        "example": "Juan Pérez",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "contacto@proveedora.com",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phone_secondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburb_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Industrial",
                        "nullable": true
                    },
                    "exterior_number": {
                        "type": "string",
                        "example": "500",
                        "nullable": true
                    },
                    "interior_number": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "address_reference": {
                        "type": "string",
                        "example": "Zona industrial norte",
                        "nullable": true
                    },
                    "rfc": {
                        "type": "string",
                        "example": "PRO850101AB2",
                        "nullable": true
                    },
                    "legal_name": {
                        "type": "string",
                        "example": "Proveedora SA de CV",
                        "nullable": true
                    },
                    "tax_system": {
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "use_cfdi": {
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "supplier_type": {
                        "type": "string",
                        "enum": [
                            "product",
                            "service",
                            "both"
                        ],
                        "example": "product"
                    },
                    "payment_terms": {
                        "type": "string",
                        "example": "30 días",
                        "nullable": true
                    },
                    "credit_limit": {
                        "type": "number",
                        "format": "float",
                        "example": 100000,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Proveedor confiable",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SuburbResource": {
                "title": "Suburb",
                "description": "Suburb resource",
                "properties": {
                    "id": {
                        "description": "ID del suburb",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del suburb",
                        "type": "string",
                        "example": "Centro"
                    },
                    "zipCodeId": {
                        "description": "ID del código postal asociado",
                        "type": "integer",
                        "example": 101
                    }
                },
                "type": "object"
            },
            "SupplierBankAccountResource": {
                "title": "SupplierBankAccount",
                "description": "Supplier bank account resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "supplierId": {
                        "type": "integer",
                        "example": 1
                    },
                    "bankName": {
                        "type": "string",
                        "example": "BBVA"
                    },
                    "accountNumber": {
                        "type": "string",
                        "example": "0123456789",
                        "nullable": true
                    },
                    "clabe": {
                        "type": "string",
                        "example": "012180001234567890",
                        "nullable": true
                    },
                    "swiftCode": {
                        "type": "string",
                        "example": "BBVAMXMMXXX",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "example": "MXN"
                    },
                    "isPrimary": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cuenta principal",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SupplierResource": {
                "title": "Supplier",
                "description": "Supplier resource with complete information",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "companyName": {
                        "type": "string",
                        "example": "Proveedora SA de CV"
                    },
                    "contactName": {
                        "type": "string",
                        "example": "Juan Pérez",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "contacto@proveedora.com",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "5512345678",
                        "nullable": true
                    },
                    "phoneSecondary": {
                        "type": "string",
                        "example": "5587654321",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "whatsapp": {
                        "type": "string",
                        "example": "5511223344",
                        "nullable": true
                    },
                    "suburbId": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Industrial",
                        "nullable": true
                    },
                    "exteriorNumber": {
                        "type": "string",
                        "example": "500",
                        "nullable": true
                    },
                    "interiorNumber": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "addressReference": {
                        "type": "string",
                        "example": "Zona industrial",
                        "nullable": true
                    },
                    "fullAddress": {
                        "type": "string",
                        "example": "Av. Industrial 500, Int. A, Col. Centro",
                        "nullable": true
                    },
                    "rfc": {
                        "type": "string",
                        "example": "PRO850101AB2",
                        "nullable": true
                    },
                    "legalName": {
                        "type": "string",
                        "example": "Proveedora SA de CV",
                        "nullable": true
                    },
                    "taxSystem": {
                        "type": "string",
                        "example": "601",
                        "nullable": true
                    },
                    "useCfdi": {
                        "type": "string",
                        "example": "G03",
                        "nullable": true
                    },
                    "supplierType": {
                        "type": "string",
                        "enum": [
                            "product",
                            "service",
                            "both"
                        ],
                        "example": "product"
                    },
                    "paymentTerms": {
                        "type": "string",
                        "example": "30 días",
                        "nullable": true
                    },
                    "creditLimit": {
                        "type": "number",
                        "format": "float",
                        "example": 100000,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended",
                            "blacklisted"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Proveedor confiable",
                        "nullable": true
                    },
                    "createdBy": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "updatedBy": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "deletedBy": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "suburb": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SuburbResource"
                            }
                        ],
                        "nullable": true
                    },
                    "bankAccounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SupplierBankAccountResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}